Django/django-rest-framework/requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
Faker==18.6.0
Markdown==3.4.3
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
validate-docbr==1.10.0
